feat: add request for all products
