+++ react-rss/src/api/types.ts
@@ -1,0 +1,11 @@
+export interface Product {
+  id: number
+  title: string
+  price: number
+  brand: string
+  rating: number
+  stock: number
+  description: string
+  category: string
+  image: string[]
+}
