+++ react-rss/src/pages/HomePage/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './useFetchAllProducts'
 export * from './hooks'
